Fix BeforeDestroyed and Destroyed event firing for request scope.
Added test, renamed test package to align it with other tests.
